requirements.txt
mss
pyperclip
pillow
pyzbar
opencv-python
numpy
